requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.8
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
